requirements.txt
asgiref==3.7.1
certifi==2023.5.7
charset-normalizer==3.1.0
Django==3.2.19
django-filter==23.2
django-request==1.6.2
djangorestframework==3.14.0
idna==3.4
psycopg2-binary==2.9.6
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.2
urllib3==2.0.2
